src/pyproject.toml
[project]
name = "mini-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles==25.1.0",
    "fastapi==0.135.2",
    "langchain==1.2.15",
    "langchain-community==0.4.1",
    "langchain-text-splitters==1.1.1",
    "motor==3.7.1",
    "pydantic-settings==2.13.1",
    "pymupdf==1.27.2.2",
    "python-dotenv==1.2.2",
    "python-multipart==0.0.22",
    "uvicorn[standard]==0.42.0",
]
